7th

+++ src/com/game/Trail.java
@@ -4,14 +4,14 @@ import java.awt.*;
 
 public class Trail extends GameObject {
 
-    private float alpha =1;
+    private float alpha = 1;
     private Handler handler;
     private Color color;
-    private int width;
-    private int height;
+    private float width;
+    private float height;
     private float life;
 
-    public Trail(int x, int y, ID id, Handler handler, Color color, int width, int height, float life) {
+    public Trail(float x, float y, ID id, Handler handler, Color color, int width, int height, float life) {
         super(x, y, id);
         this.handler = handler;
         this.color = color;
@@ -21,10 +21,10 @@ public class Trail extends GameObject {
     }
 
     public void tick() {
-        if (alpha > life){
-            alpha -=life -0.0001f;
+        if (alpha > life) {
+            alpha -= life - 0.0001f;
 
-        }else handler.removeObject(this);
+        } else handler.removeObject(this);
 
     }
 
@@ -34,11 +34,9 @@ public class Trail extends GameObject {
         g2d.setComposite(makeTransparent(alpha));
 
         g.setColor(color);
-        g.fillRect(x, y, width-2, height-2);
+        g.fillRect((int) x, (int) y, (int) width - 2, (int) height - 2);
 
         g2d.setComposite(makeTransparent(1));
-
-
     }
 
     private AlphaComposite makeTransparent(float alpha) {
